pyproject.toml
[project]
name = "codepush-mcp-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.4",
    "fastapi>=0.116.1",
    "fastmcp>=2.10.5",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.3",
]
